expose `useHandleLibrary` hook
